Update to v0.5.0
Modified the flow to be more failsafe. If an error occurs the
the images are normally still present.

Added the backup only capability (-b).

Improved messaging.

Added the -k option to force the images to always be kept when
doing the resize operation. It combines the resize and backup
operations into one.